campsiteapp/requirements.txt
Django # Django baze
djangorestframework # drf API
psycopg2-binary # для psql
mozilla-django-oidc # для keycloak
requests # базовые запросы
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.2
Django==5.2.1
django-allauth==65.7.0
django-keycloak-auth==1.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
josepy==2.0.0
jwt
mozilla-django-oidc==4.0.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
locust==2.37.1
django-cors-headers
Pillow
